+++ lib/utils/reset.dart
@@ -1,6 +1,6 @@
 // Reset the app
 //
-// Time-stamp: <Thursday 2024-06-06 05:58:50 +1000 Graham Williams>
+// Time-stamp: <Saturday 2024-06-15 14:33:18 +1000 Graham Williams>
 //
 /// Copyright (C) 2024, Togaware Pty Ltd
 ///
@@ -29,17 +29,17 @@ library;
 
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:rattle/provider/dataset_loaded.dart';
-import 'package:rattle/provider/path.dart';
-import 'package:rattle/provider/stdout.dart';
-import 'package:rattle/provider/terminal.dart';
-import 'package:rattle/provider/wordcloud/build.dart';
-import 'package:rattle/provider/wordcloud/checkbox.dart';
-import 'package:rattle/provider/wordcloud/maxword.dart';
-import 'package:rattle/provider/wordcloud/minfreq.dart';
-import 'package:rattle/provider/wordcloud/punctuation.dart';
-import 'package:rattle/provider/wordcloud/stem.dart';
-import 'package:rattle/provider/wordcloud/stopword.dart';
+import 'package:rattle/providers/dataset_loaded.dart';
+import 'package:rattle/providers/path.dart';
+import 'package:rattle/providers/stdout.dart';
+import 'package:rattle/providers/terminal.dart';
+import 'package:rattle/providers/wordcloud/build.dart';
+import 'package:rattle/providers/wordcloud/checkbox.dart';
+import 'package:rattle/providers/wordcloud/maxword.dart';
+import 'package:rattle/providers/wordcloud/minfreq.dart';
+import 'package:rattle/providers/wordcloud/punctuation.dart';
+import 'package:rattle/providers/wordcloud/stem.dart';
+import 'package:rattle/providers/wordcloud/stopword.dart';
 import 'package:rattle/r/start.dart';
 import 'package:xterm/xterm.dart';
 
@@ -61,7 +61,7 @@ void reset(BuildContext context, WidgetRef ref) {
   ref.read(stdoutProvider.notifier).state = '';
   // CONSOLE TAB
   ref.read(terminalProvider.notifier).state = Terminal();
-  rStart(ref);
+  rStart(context, ref);
   // DATASET TAB
   ref.read(pathProvider.notifier).state = '';
   debugPrint('DATASET UNLOADED');